KVM: arm64: Move host EL1 code out of hyp/ directory
commit 9429f4b0412d05243237c7695c59d0a7b1174492 upstream.

kvm/hyp/reserved_mem.c contains host code executing at EL1 and is not
linked into the hypervisor object. Move the file into kvm/pkvm.c and
rework the headers so that the definitions shared between the host and
the hypervisor live in asm/kvm_pkvm.h.

Tested-by: Fuad Tabba <tabba@google.com>
Link: https://lore.kernel.org/r/20211202171048.26924-4-will@kernel.org
